Move create_person to helper file
